Backend/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==2.2
django-resized==0.3.6
django-cors-headers==3.2.1
django-smtp-ssl==1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
python-magic==0.4.18
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
zope.interface==5.1.0
